Updated Linux x86/x64 native libraries for Local class to use the kernel for obtaining graphics hardware information rather than SDL.
The returned string value by getGraphicsDevice() should now represent
the hardware rather than video driver on Linux systems.  The native code
invokes a kernel command and parses the result in order to get the
information, so depending on bash configurations on the Linux platform,
the code may or may not work (please submit an issue if it doesn't).
Tested on LinuxMint 14 KDE distro.