Mavenize paulscode soundsystem.
Paulscode soundsystem jars are not available via any remote maven
repository. Thus the only options are to download and install the jars
manually to the local maven repo, or to host a maven repository inside
of the project.

Use the latter option as it makes the build portable; i.e., anyone can
clone the project and immediately build it from source by just running
one command: mvn clean install.

- Add the project-based repository in the snap2d module under the repo
  directory.
- Add the project-based repository in the snap2d pom.
- Add the project-based repository dependencies to the snap2d pom.